e-permit-internal-api/src/main/java/epermit/controllers/KeyController.java
package epermit.controllers;

import java.util.Map;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import epermit.services.KeyService;
import lombok.RequiredArgsConstructor;

@RestController
@RequiredArgsConstructor
@RequestMapping("/keys")
public class KeyController {
    private final KeyService keyService;

    @PostMapping()
    public void create(@RequestBody Map<String, String> input) {
        keyService.create(input.get("key_id"));
    }

    @PatchMapping("{id}/enable")
    public void enable(@RequestParam Integer id) {
        keyService.enable(id);
    }
}
